refactor: 프로젝트 페이지 컴포넌트 이름 변경
